+++ src/test/java/com/vssekorin/cactoosmath/scalar/ScalarMatchTest.java
@@ -23,49 +23,36 @@
  */
 package com.vssekorin.cactoosmath.scalar;
 
-import java.util.Map;
-import org.cactoos.Func;
-import org.cactoos.func.UncheckedFunc;
+import org.cactoos.map.MapEntry;
+import org.cactoos.map.MapOf;
+import org.hamcrest.MatcherAssert;
+import org.hamcrest.Matchers;
+import org.junit.Test;
 
 /**
- * Pattern matching.
+ * Test case for {@link ScalarMatch}.
  *
  * @author Vseslav Sekorin (vssekorin@gmail.com)
  * @version $Id$
- * @param <X> Type of input.
- * @param <Y> Type of output.
  * @since 0.2
+ * @checkstyle JavadocMethodCheck (500 lines)
+ * @checkstyle MagicNumberCheck (500 lines)
  */
-public final class Match<X, Y> implements Func<X, Y> {
+public final class ScalarMatchTest {
 
-    /**
-     * Map.
-     */
-    private final Map<Func<X, Boolean>, Func<X, Y>> map;
-
-    /**
-     * Default case.
-     */
-    private final Func<X, Y> other;
-
-    /**
-     * Ctor.
-     * @param src Map
-     * @param dflt Default case
-     */
-    public Match(final Map<Func<X, Boolean>, Func<X, Y>> src,
-        final Func<X, Y> dflt) {
-        this.map = src;
-        this.other = dflt;
-    }
-
-    @Override
-    public Y apply(final X input) throws Exception {
-        return this.map.entrySet().stream()
-            .filter(item -> new UncheckedFunc<>(item.getKey()).apply(input))
-            .map(Map.Entry::getValue)
-            .findFirst()
-            .orElse(this.other)
-            .apply(input);
+    @Test
+    public void apply() throws Exception {
+        final int fst = 3;
+        final int snd = 6;
+        MatcherAssert.assertThat(
+            new ScalarMatch<>(
+                new MapOf<>(
+                    new MapEntry<>(() -> fst < snd, () -> 1),
+                    new MapEntry<>(() -> snd < fst, () -> -1)
+                ),
+                () -> 0
+            ),
+            Matchers.equalTo(8)
+        );
     }
 }
